+++ src/gpu/Swizzle.h
@@ -38,7 +38,7 @@ public:
 
     constexpr char operator[](int i) const {
         SkASSERT(i >= 0 && i < 4);
-        int idx = (fKey >> (4U * i)) & 0xfU;
+        int idx = (fKey >> (4 * i)) & 0xfU;
         return IToC(idx);
     }
 
@@ -62,7 +62,7 @@ public:
 private:
     explicit constexpr Swizzle(uint16_t key) : fKey(key) {}
 
-    static constexpr float ComponentIndexToFloat(std::array<float, 4>, int idx);
+    static constexpr float ComponentIndexToFloat(std::array<float, 4>, size_t idx);
     static constexpr int CToI(char c);
     static constexpr char IToC(int idx);
 
@@ -70,7 +70,8 @@ private:
 };
 
 constexpr Swizzle::Swizzle(const char c[4])
-        : fKey((CToI(c[0]) << 0) | (CToI(c[1]) << 4) | (CToI(c[2]) << 8) | (CToI(c[3]) << 12)) {}
+        : fKey(static_cast<uint16_t>((CToI(c[0]) << 0) | (CToI(c[1]) << 4) | (CToI(c[2]) << 8) |
+                                     (CToI(c[3]) << 12))) {}
 
 constexpr Swizzle::Swizzle(const Swizzle& that)
         : fKey(that.fKey) {}
@@ -83,7 +84,7 @@ constexpr Swizzle& Swizzle::operator=(const Swizzle& that) {
 constexpr std::array<float, 4> Swizzle::applyTo(std::array<float, 4> color) const {
     uint32_t key = fKey;
     // Index of the input color that should be mapped to output r.
-    int idx = (key & 15);
+    size_t idx = (key & 15);
     float outR = ComponentIndexToFloat(color, idx);
     key >>= 4;
     idx = (key & 15);
@@ -97,14 +98,14 @@ constexpr std::array<float, 4> Swizzle::applyTo(std::array<float, 4> color) cons
     return { outR, outG, outB, outA };
 }
 
-constexpr float Swizzle::ComponentIndexToFloat(std::array<float, 4> color, int idx) {
+constexpr float Swizzle::ComponentIndexToFloat(std::array<float, 4> color, size_t idx) {
     if (idx <= 3) {
         return color[idx];
     }
-    if (idx == CToI('1')) {
+    if (idx == static_cast<size_t>(CToI('1'))) {
         return 1.0f;
     }
-    if (idx == CToI('0')) {
+    if (idx == static_cast<size_t>(CToI('0'))) {
         return 0.0f;
     }
     SkUNREACHABLE;
@@ -138,11 +139,11 @@ constexpr char Swizzle::IToC(int idx) {
 constexpr Swizzle Swizzle::Concat(const Swizzle& a, const Swizzle& b) {
     uint16_t key = 0;
     for (int i = 0; i < 4; ++i) {
-        int idx = (b.fKey >> (4U * i)) & 0xfU;
+        int idx = (b.fKey >> (4 * i)) & 0xfU;
         if (idx != CToI('0') && idx != CToI('1')) {
             SkASSERT(idx >= 0 && idx < 4);
             // Get the index value stored in a at location idx.
-            idx = ((a.fKey >> (4U * idx)) & 0xfU);
+            idx = ((a.fKey >> (4 * idx)) & 0xfU);
         }
         key |= (idx << (4U * i));
     }
